mGear -replaced intensity float variables for 16-bit unsigned integers in the lighting functions
mEngineer
-replaced intensity float variables for 16-bit unsigned integers in the lighting functions
-fixed remove lightmaps bug
-fixed changing the Spot angle value changing temporary view of the lightmap to the spotlight type
-fixed mouse scroll not changing light falloff during lightmap editing
-fixed being able to create another lightmap for a object which already have one

+++ mGear-1/mGear-1/engine.c
@@ -444,7 +444,7 @@ uint8 FillAlphaLight(unsigned char *data, uint8 r, uint8 g, uint8 b, uint16 w, u
 	return 1;
 }
 
-uint32 AddLightToAlphaLight(unsigned char *data, uint16 w, uint16 h, uint8 r, uint8 g, uint8 b, float falloff, uint16 x, uint16 y, uint16 z, float intensity, LIGHT_TYPE type)
+uint32 AddLightToAlphaLight(unsigned char *data, uint16 w, uint16 h, uint8 r, uint8 g, uint8 b, float falloff, uint16 x, uint16 y, uint16 z, uint16 intensity, LIGHT_TYPE type)
 {
 	register uint16 i, j;
 	float d, att;
@@ -517,7 +517,7 @@ uint32 AddLightToAlphaLight(unsigned char *data, uint16 w, uint16 h, uint8 r, ui
 	return 1;
 }
 
-uint32 AddSpotlightToAlphaLight(unsigned char *data, uint16 w, uint16 h, uint8 r, uint8 g, uint8 b, float falloff, uint16 x, uint16 y, uint16 z, float intensity, LIGHT_TYPE type, uint16 x2, uint16 y2, uint16 ang)
+uint32 AddSpotlightToAlphaLight(unsigned char *data, uint16 w, uint16 h, uint8 r, uint8 g, uint8 b, float falloff, uint16 x, uint16 y, uint16 z, uint16 intensity, LIGHT_TYPE type, uint16 x2, uint16 y2, uint16 ang)
 {
 	register uint16 i, j;
 	float d, att;
@@ -668,7 +668,7 @@ uint8 FillLightmap(unsigned char *data, uint8 r, uint8 g, uint8 b, uint16 w, uin
 	return 1;
 }
 
-uint32 AddLightToLightmap(unsigned char *data, uint16 w, uint16 h, uint8 r, uint8 g, uint8 b, float falloff, uint16 x, uint16 y, uint16 z, float intensity, LIGHT_TYPE type)
+uint32 AddLightToLightmap(unsigned char *data, uint16 w, uint16 h, uint8 r, uint8 g, uint8 b, float falloff, uint16 x, uint16 y, uint16 z, uint16 intensity, LIGHT_TYPE type)
 {
 	register uint16 i, j;
 	float d, att;
@@ -724,7 +724,7 @@ uint32 AddLightToLightmap(unsigned char *data, uint16 w, uint16 h, uint8 r, uint
 	return 1;
 }
 
-uint32 AddSpotlightToLightmap(unsigned char *data, uint16 w, uint16 h, uint8 r, uint8 g, uint8 b, float falloff, uint16 x, uint16 y, uint16 z, float intensity, LIGHT_TYPE type, uint16 x2, uint16 y2, uint16 ang)
+uint32 AddSpotlightToLightmap(unsigned char *data, uint16 w, uint16 h, uint8 r, uint8 g, uint8 b, float falloff, uint16 x, uint16 y, uint16 z, uint16 intensity, LIGHT_TYPE type, uint16 x2, uint16 y2, uint16 ang)
 {
 	register  i, j;
 	float d, att;
